Update how CogAt weighting is performed.
Before, circular relationships in the Cognitive Atlas would cause the
weighting to get stuck in an infinite loop. Now, we define the full
chain of relationships in the weights for a single term, so that the
dot product is only performed once.

Also, make some changes to data preparation and the CV prototype script.